refactor: move route into config
